fixed unblock user issue + start working on login wrong credentials

+++ client/src/app/services/login.service.ts
@@ -29,23 +29,25 @@ export class LoginService {
       tap( (data) => {
         //Otteniamo la response dal server (vedere route login.js in server per capire com'è fatta la response)
         this.response = data;
-        if(this.response != 'banned'){
-          if(this.response.needspasswordchange != undefined &&
-             this.response.needspasswordchange != null &&
-             this.response.needspasswordchange == true){
-            localStorage.clear()
-            localStorage.setItem('username', JSON.stringify(this.response.username))
-          }
-          else{
-            console.log(this.response)
-            localStorage.clear()
-            //Salviamo in localstorage il token
-            this.token = this.response.token;
-            localStorage.setItem('auth_token', this.token);
+        if(this.response != 'error'){
+          if(this.response != 'banned'){
+            if(this.response.needspasswordchange != undefined &&
+              this.response.needspasswordchange != null &&
+              this.response.needspasswordchange == true){
+              localStorage.clear()
+              localStorage.setItem('username', JSON.stringify(this.response.username))
+            }
+            else{
+              console.log(this.response)
+              localStorage.clear()
+              //Salviamo in localstorage il token
+              this.token = this.response.token;
+              localStorage.setItem('auth_token', this.token);
 
-            //Salviamo in localstorage l'utente per averne i dati comodamente estraibili in seugito
-            this.user = this.response.user;
-            localStorage.setItem('current_user', JSON.stringify(this.user));
+              //Salviamo in localstorage l'utente per averne i dati comodamente estraibili in seugito
+              this.user = this.response.user;
+              localStorage.setItem('current_user', JSON.stringify(this.user));
+            }
           }
         }
       }));
